+++ Magic/CHANGELOG.md
@@ -8,6 +8,7 @@
  - Fix world gen errors by reducing max world height by one
  - More Valhalla integration:
    - Combine profile and skill XP so there is just one read/write representation of both
+   - The level attributes are now keyed by skill instead of profile
    - Force update wand lore on level up, in case mana is driven by Valhalla
    - Add a `magic` profile and associated skill, can be disabled via config